+++ Source/Plugins/Reconstruction/ReconstructionFilters/ScalarSegmentFeatures.cpp
@@ -143,7 +143,7 @@ class TSpecificCompareFunctor : public CompareFunctor
 
   private:
     T* m_Data; // The data that is being compared
-    size_t m_Length; // Length of the Data Array
+    int64_t m_Length; // Length of the Data Array
     T      m_Tolerance; // The tolerance of the comparison
     int32_t* m_FeatureIds; // The Feature Ids
 };
